Separates concerns into modules for easier testing
Uses ES5 Syntax
