+++ modules/githubAuth.js
@@ -1,0 +1,19 @@
+var GitHubApi = require("github");
+
+var github = new GitHubApi({
+    version: "3.0.0",
+    debug: true,
+    protocol: "https",
+    host: "api.github.com",
+    timeout: 5000,
+    headers: {
+        "user-agent": "Hacktoberfest Checker"
+    }
+});
+
+github.authenticate({
+    type: "oauth",
+    token: process.env.GITHUB_TOKEN
+});
+
+module.exports = github;
